# Improve vectorization of IndexOf(chars, StringComparison.OrdinalIgnoreCase)