Use the same general "Algorithm 1: Generic SIMD" that we do for StringComparison.Ordinal, adapted for OrdinalIgnoreCase.

```C#
private static readonly string s_haystack = new HttpClient().GetStringAsync("https://www.gutenberg.org/files/1661/1661-0.txt").Result;

[Params("watson", "elementary", "holmes", "the")]
public string Needle { get; set; }

[Benchmark]
public int Count()
{
    int count = 0;
    ReadOnlySpan<char> haystack = s_haystack;
    while (true)
    {
        int pos = haystack.IndexOf(Needle, StringComparison.OrdinalIgnoreCase);
        if (pos < 0) break;
        count++;
        haystack = haystack.Slice(pos + Needle.Length);
    }
    return count;
}
```

| Method |         Toolchain |     Needle |      Mean |    Error |   StdDev | Ratio |
|------- |------------------ |----------- |----------:|---------:|---------:|------:|
|  Count | \main\corerun.exe | elementary | 580.54 us | 3.562 us | 2.781 us |  1.00 |
|  Count |   \pr\corerun.exe | elementary |  59.37 us | 0.447 us | 0.397 us |  0.10 |
|        |                   |            |           |          |          |       |
|  Count | \main\corerun.exe |     holmes | 366.58 us | 0.607 us | 0.568 us |  1.00 |
|  Count |   \pr\corerun.exe |     holmes |  82.55 us | 0.204 us | 0.181 us |  0.23 |
|        |                   |            |           |          |          |       |
|  Count | \main\corerun.exe |        the | 547.91 us | 1.257 us | 1.050 us |  1.00 |
|  Count |   \pr\corerun.exe |        the | 258.62 us | 1.123 us | 0.996 us |  0.47 |
|        |                   |            |           |          |          |       |
|  Count | \main\corerun.exe |     watson | 230.76 us | 0.550 us | 0.514 us |  1.00 |
|  Count |   \pr\corerun.exe |     watson |  58.24 us | 0.448 us | 0.419 us |  0.25 |